requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uuid==1.30
uvicorn==0.32.0
